Fix linters and add Quote folder and file and also fix calculator links
