+++ src/App.css
@@ -11,11 +11,8 @@ body {
 
 .calculator {
   border: 1px solid #ccc;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 500px;
+  left: 100vh;
+  width: 40%;
 }
 
 .calculator-screen {
