+++ src/Components/Calculator.js
@@ -10,42 +10,42 @@ const CalculatorHook = () => {
   };
 
   return (
+    <>
+      <h3 className="mt-3">Lets Do Some Math !!</h3>
+      <div className="calculator card">
 
-    <div className="calculator card">
+        <div type="text" className="calculator-screen z-depth-1 " placeholder="0" disabled>
+          {stateObj.total}
+          {stateObj.operation}
+          {stateObj.next}
+        </div>
+        <div className="calculator-keys">
 
-      <div type="text" className="calculator-screen z-depth-1 " placeholder="0" disabled>
-        {stateObj.total}
-        {stateObj.operation}
-        {stateObj.next}
+          <button type="button" onClick={onClickHandler} className="operator btn btn-gray border-secondary">+/-</button>
+          <button type="button" onClick={onClickHandler} className="operator btn btn-gray border-secondary">AC</button>
+          <button type="button" onClick={onClickHandler} className="operator btn btn-gray border-secondary">%</button>
+          <button type="button" onClick={onClickHandler} className="operator btn btn-Orange border-secondary">&divide;</button>
+          <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">7</button>
+          <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">8</button>
+          <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">9</button>
+          <button type="button" onClick={onClickHandler} className="operator btn btn-Orange border-secondary">+</button>
+
+          <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">4</button>
+          <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">5</button>
+          <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">6</button>
+          <button type="button" onClick={onClickHandler} className="operator btn btn-Orange border-secondary">-</button>
+
+          <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">1</button>
+          <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">2</button>
+          <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">3</button>
+          <button type="button" onClick={onClickHandler} className="operator btn btn-Orange border-secondary">x</button>
+
+          <button type="button" onClick={onClickHandler} className="btn btn-gray zero border-secondary">0</button>
+          <button type="button" onClick={onClickHandler} className="decimal function btn btn-gray dot border-secondary">.</button>
+          <button type="button" onClick={onClickHandler} className="equal-sign operator btn btn-Orange equal border-secondary">=</button>
+        </div>
       </div>
-      <div className="calculator-keys">
-
-        <button type="button" onClick={onClickHandler} className="operator btn btn-gray border-secondary">+/-</button>
-        <button type="button" onClick={onClickHandler} className="operator btn btn-gray border-secondary">AC</button>
-        <button type="button" onClick={onClickHandler} className="operator btn btn-gray border-secondary">%</button>
-        <button type="button" onClick={onClickHandler} className="operator btn btn-Orange border-secondary">&divide;</button>
-
-        <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">7</button>
-        <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">8</button>
-        <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">9</button>
-        <button type="button" onClick={onClickHandler} className="operator btn btn-Orange border-secondary">+</button>
-
-        <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">4</button>
-        <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">5</button>
-        <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">6</button>
-        <button type="button" onClick={onClickHandler} className="operator btn btn-Orange border-secondary">-</button>
-
-        <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">1</button>
-        <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">2</button>
-        <button type="button" onClick={onClickHandler} className="btn btn-gray waves-effect border-secondary">3</button>
-        <button type="button" onClick={onClickHandler} className="operator btn btn-Orange border-secondary">x</button>
-
-        <button type="button" onClick={onClickHandler} className="btn btn-gray zero border-secondary">0</button>
-        <button type="button" onClick={onClickHandler} className="decimal function btn btn-gray dot border-secondary">.</button>
-        <button type="button" onClick={onClickHandler} className="equal-sign operator btn btn-Orange equal border-secondary">=</button>
-
-      </div>
-    </div>
+    </>
   );
 };
