+++ src/Components/Navbar/Navbar.js
@@ -4,11 +4,11 @@ import { Link } from 'react-router-dom';
 function Navbar() {
   return (
     <nav className="p-2 navbar d-flex justify-content-between navbar-light">
-      <span className="title-math">Math Magicians</span>
+      <Link to="/" className="title-math text-black text-decoration-none">Math Magicians</Link>
       <div className="menu-items">
-        <Link to="/math-magician-app/Calculator">Calculator</Link>
-        <Link to="/math-magicians/" className="links text-decoration-none">Home</Link>
-        <Link to="/math-magicians/Calculator" className="links text-decoration-none">Calculator</Link>
+        <Link to="/" className="links text-decoration-none">Home</Link>
+        <Link to="/Calculator" className="links text-decoration-none">Calculator</Link>
+        <Link to="/Quote" className="links text-decoration-none">Quote</Link>
       </div>
     </nav>
   );
