+++ src/Components/Quote/Quote.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+function Quote() {
+  return (
+    <div className="m-5">
+      <div className="m-5 paragraph">
+        <p>
+          You all need is to calculate life...!
+        </p>
+      </div>
+    </div>
+  );
+}
+
+export default Quote;
