--- src/Home/Home.js
@@ -1,20 +1,0 @@
-mport React from 'react';
-
-function Home() {
-  return (
-    <div className="m-5">
-      <h1 className="mb-5">Welcome to our page!</h1>
-      <p>
-      Perform four fundamental operations and engineering calculations. 
-      To start the engineering calculator, tap the engineering calculator icon.
-      </p>
-      <br />
-      <p>
-      We recognize that our users may have various Internet Browsers and Operating Systems or devices. We like our visitors to have the best possible experience when using our applications.
-      The Virginia Workers' Compensation Commission recommends Google Chrome for best results or any other modern web browsers with HTML5 and CSS3 compatibility.
-      </p>
-    </div>
-  );
-}
-
-export default Home;
